pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "crewai>=0.30.11",
    "flask>=3.1.1",
    "langchain>=0.1.20",
    "langchain-anthropic>=0.1.13",
    "langchain-groq>=0.1.5",
    "langchain-mistralai>=0.1.7",
    "langchain-openai>=0.1.7",
    "openai>=1.82.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]
